Send the whole patch to device on AMP cab id param change
In Multistomp patches, AMP effect's cab id parameter value is actually
an effect id of the corresponding AMP effect. However, the Parameter
Change SysEx (0x31) allows only two 7-bit bytes to encode the value.
Thus, only the lower byte of the cab id is passed, the remaining
bytes are assumed the same (corresponding to Multistomp model) and
are filled in by the fw.

To keep the patch data consistent, it's preferred to send the whole
patch which includes the correct/full value for the cab id.