Scanning can be done in read and update mode
The scanning is done sequentially for all patches (1 to 50).
Each patch should first be explicitly set (MIDI 0xc0, ProgramChange).

- read mode: requests the current patch data (SysEx: 0x29);
- update mode: first sends the editor's data for the current patch,
  (SysEx: 0x28), then requests the current patch data (SysEx: 0x29)

With update mode, if any of effects referenced in editor's patch data
are not actually installed on the device, the patch recevied back from
the device will have these effects set to THRU.