Add a control knob for editing BPM parameter of the current patch
The updated BPM value is sent to the device as part of the whole
patch data via SysEx:0x28.

NOTE: The pedal fw only supports TAP function. It does not send
the updated BPM value (in the patch data) to the editor; it just
updates the patch internally. So user should avoid using the TAP
function while attached to the editor. Otherwise, the editor data
will override the BPM value set via TAP.